move builder works to dev-builder.ts
