+++ src/commands/dev/lib/dev-server.ts
@@ -1,35 +1,27 @@
-import fs from 'fs';
-import path from 'path';
 import http from 'http';
-
 import chalk from 'chalk';
-import ignore from 'ignore';
 import serveHandler from 'serve-handler';
-import glob from '@now/build-utils/fs/glob';
-import FileFsRef from '@now/build-utils/file-fs-ref';
 // @ts-ignore
 import { createFunction } from '../../../../../lambdas/lambda-dev/dist/src';
 
-import wait from '../../../util/output/wait';
 import error from '../../../util/output/error';
 import success from '../../../util/output/success';
-import { NowError } from '../../../util/now-error';
 import { readLocalConfig } from '../../../util/config/files';
 
-import builderCache from './builder-cache';
 import devRouter from './dev-router';
+import { buildUserProject, createIgnoreList } from './dev-builder';
 
 import {
   DevServerStatus,
   DevServerOptions,
-  BuildConfig,
   BuilderOutput,
   BuilderOutputs,
   HttpHandler
 } from './types';
 
 export default class DevServer {
-  private cwd: string;
+  public cwd: string;
+
   private debug: boolean;
   private server: http.Server;
   private status: DevServerStatus;
@@ -96,7 +88,7 @@ export default class DevServer {
         // Initial build. Not meant to invoke, but for speed up further builds.
         if (nowJson && nowJson.builds) {
           this.logDebug(`Initial build`);
-          await this.buildUserProject(nowJson.builds, this.cwd);
+          await buildUserProject(nowJson.builds, this);
         }
 
         this.logSuccess('Initial build ready');
@@ -150,9 +142,9 @@ export default class DevServer {
     cwd: string
   ) => {
     const filePath = req.url ? req.url.replace(/^\//, '') : '';
-    const ig = createIgnoreList(cwd);
+    const ignore = createIgnoreList(cwd);
 
-    if (filePath && ig.ignores(filePath)) {
+    if (filePath && ignore.ignores(filePath)) {
       res.writeHead(404);
       res.end();
       return;
@@ -191,7 +183,8 @@ export default class DevServer {
 
     // build source files to assets
     this.logDebug('Start builders', nowJson.builds);
-    const assets = await this.buildUserProject(nowJson.builds, cwd);
+    const assets = await buildUserProject(nowJson.builds, this);
+
     this.logDebug('Built', Object.keys(assets));
 
     // find asset responsible for dest
@@ -235,102 +228,6 @@ export default class DevServer {
         return res.end();
     }
   };
-
-  /**
-   * Build project to statics & lambdas
-   */
-  buildUserProject = async (buildsConfig: BuildConfig[], cwd: string) => {
-    try {
-      this.setStatusBusy('installing builders');
-      await this.installBuilders(buildsConfig);
-
-      this.setStatusBusy('building lambdas');
-      const assets = await this.buildLambdas(buildsConfig, cwd);
-
-      this.setStatusIdle();
-      return assets;
-    } catch (err) {
-      this.setStatusIdle();
-      throw new Error('Build failed.');
-    }
-  };
-
-  installBuilders = async (buildsConfig: BuildConfig[]) => {
-    const builders = buildsConfig
-      .map(build => build.use)
-      .filter(pkg => pkg !== '@now/static')
-      .concat('@now/build-utils');
-
-    for (const builder of builders) {
-      const stopSpinner = wait(`installing ${builder}`);
-      await builderCache.install(builder);
-      stopSpinner();
-    }
-  };
-
-  buildLambdas = async (buildsConfig: BuildConfig[], cwd: string) => {
-    const ignores = createIgnoreList(cwd);
-    const files = await collectProjectFiles('**', cwd, ignores);
-    let results = {};
-
-    for (const build of buildsConfig) {
-      try {
-        this.logDebug(`Build ${JSON.stringify(build)}`);
-
-        const builder = builderCache.get(build.use);
-
-        const entries = Object.values(
-          await collectProjectFiles(build.src, cwd, ignores)
-        );
-
-        for (const entry of entries) {
-          const output = await builder.build({
-            files,
-            entrypoint: path.relative(cwd, entry.fsPath),
-            workPath: cwd,
-            config: build.config
-          });
-          results = { ...results, ...output };
-        }
-      } catch (err) {
-        throw new NowError({
-          code: 'NOW_BUILDER_FAILURE',
-          message: `Failed building ${chalk.bold(build.src)} with ${build.use}`,
-          meta: err.stack
-        });
-      }
-    }
-
-    return results;
-  };
-}
-
-/**
- * Create ignore list according .gitignore & .nowignore in cwd
- */
-function createIgnoreList(cwd: string) {
-  const ig = ignore();
-
-  const gitignore = path.join(cwd, '.gitignore');
-  const nowignore = path.join(cwd, '.nowignore');
-
-  if (fs.existsSync(gitignore)) {
-    ig.add(fs.readFileSync(gitignore, 'utf8'));
-  }
-
-  if (fs.existsSync(nowignore)) {
-    ig.add(fs.readFileSync(nowignore, 'utf8'));
-  }
-
-  // special case for now-cli's usage
-  ig.add('.nowignore');
-
-  // temp workround for excluding ncc/ & user/ folder generated by builders
-  // should be removed later.
-  ig.add('ncc');
-  ig.add('user');
-
-  return ig;
 }
 
 /**
@@ -345,22 +242,6 @@ function serveStaticFile (
     public: cwd,
     cleanUrls: false
   })
-}
-
-/**
- * Collect project files, with .gitignore and .nowignore honored.
- */
-async function collectProjectFiles(pattern: string, cwd: string, ignore: any) {
-  const files = await glob(pattern, cwd);
-  const filteredFiles: { [key: string]: FileFsRef } = {};
-
-  Object.entries(files).forEach(([name, file]) => {
-    if (!ignore.ignores(name)) {
-      filteredFiles[name] = file;
-    }
-  });
-
-  return filteredFiles;
 }
 
 /**
